LPS-102145 Move comparator to where it is used
