+++ modules/apps/change-tracking/change-tracking-api/bnd.bnd
@@ -11,5 +11,4 @@ Export-Package:\
 	com.liferay.change.tracking.function,\
 	com.liferay.change.tracking.model,\
 	com.liferay.change.tracking.service,\
-	com.liferay.change.tracking.service.persistence,\
-	com.liferay.change.tracking.util.comparator
+	com.liferay.change.tracking.service.persistence
