--- modules/apps/change-tracking/change-tracking-api/src/main/resources/com/liferay/change/tracking/util/comparator/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
