+++ portal-impl/src/portal.properties
@@ -7265,7 +7265,6 @@
         com.liferay.calendar.util.comparator.CalendarNameComparator,\
         com.liferay.calendar.util.comparator.CalendarResourceCodeComparator,\
         com.liferay.calendar.util.comparator.CalendarResourceNameComparator,\
-        com.liferay.change.tracking.util.comparator.CTEntryCreateDateComparator,\
         com.liferay.document.library.content.util.comparator.DLContentVersionComparator,\
         com.liferay.document.library.file.rank.util.comparator.FileRankCreateDateComparator,\
         com.liferay.document.library.kernel.util.comparator.FolderIdComparator,\
